fix(stream): fix go-jsonnet type compatibility in OTLP flatten code
Convert all uint32/uint64/typed-slice fields to go-jsonnet-compatible
types (int64/[]any) to prevent "Not a json type" errors at record write time.
Also add empty-batch source validation so wrong-secret requests get 404
instead of a free 200 on empty OTLP payloads.

+++ internal/pkg/service/stream/source/type/otlpsource/flatten_traces.go
@@ -52,7 +52,7 @@ func flattenSpan(
 	rec.Set("trace_state", span.TraceState().AsRaw())
 	rec.Set("name", span.Name())
 	rec.Set("kind", span.Kind().String())
-	rec.Set("flags", span.Flags())
+	rec.Set("flags", int64(span.Flags()))
 	rec.Set("status_code", span.Status().Code().String())
 	rec.Set("status_message", span.Status().Message())
 	rec.Set("attributes", attributesToMap(span.Attributes()))
